chore: move message logic to lib dir
